requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pdf2image>=1.16.3
langchain>=0.0.335
pillow>=10.1.0
python-dotenv>=1.0.0
pydantic>=2.4.2
google-generativeai
google-cloud-vision
requests>=2.31.0
aiohttp>=3.9.0  
boto3>=1.26.0  